rename stacked to planar
