+++ benches/blocks.rs
@@ -1,4 +1,4 @@
-use audio_blocks::{interleaved::Interleaved, ops::Ops, sequential::Sequential, stacked::Stacked};
+use audio_blocks::{interleaved::Interleaved, ops::Ops, planar::Planar, sequential::Sequential};
 use criterion::{Criterion, criterion_group, criterion_main};
 
 pub fn bench_three_types(c: &mut Criterion, num_channels: u16, num_frames: usize) {
@@ -14,9 +14,9 @@ pub fn bench_three_types(c: &mut Criterion, num_channels: u16, num_frames: usize
         |b| b.iter(|| block.for_each(|v| *v *= 2.0)),
     );
 
-    let mut block = Stacked::<f32>::new(num_channels, num_frames);
+    let mut block = Planar::<f32>::new(num_channels, num_frames);
     c.bench_function(
-        &format!("for each stacked {num_channels}ch {num_frames}fr"),
+        &format!("for each planar {num_channels}ch {num_frames}fr"),
         |b| b.iter(|| block.for_each(|v| *v *= 2.0)),
     );
 }
