+++ src/planar/mod.rs
@@ -4,6 +4,6 @@ mod view;
 mod view_mut;
 
 #[cfg(any(feature = "std", feature = "alloc"))]
-pub use owned::Stacked;
-pub use view::{StackedPtrAdapter, StackedView};
-pub use view_mut::{StackedPtrAdapterMut, StackedViewMut};
+pub use owned::Planar;
+pub use view::{PlanarPtrAdapter, PlanarView};
+pub use view_mut::{PlanarPtrAdapterMut, PlanarViewMut};
